Feat: 당첨 내역을 출력한다.

+++ src/main/java/lotto/util/DecimalFormatter.java
@@ -10,7 +10,7 @@ public class DecimalFormatter {
 
     }
 
-    public static String toNumberWithDelimiter(int numberInput) {
+    public static String toNumberWithDelimiter(long numberInput) {
         return decimalFormat.format(numberInput);
     }
 }
